Add a SystemPlugins dll/proj (for now retroeditor does not compile this. Refactor things to use the new library
