+++ Plugins/Plugins.csproj
@@ -5,6 +5,7 @@
   </ItemGroup>
 
   <ItemGroup Condition="'$(RetroEditorBuilding)' == 'Yes'">
+    <ProjectReference Include="..\SystemPlugins\SystemPlugins.csproj" />
     <ProjectReference Include="..\RetroEditor\RetroEditor.csproj" />
   </ItemGroup>
