+++ Plugins/Source/GamePlugins/LegendOfZeldaLinkToThePast/LinkToThePast.cs
