+++ Plugins/Source/GamePlugins/SuperMarioWorld/SuperMarioWorld.cs
@@ -7,11 +7,12 @@ using System;
 
 using RetroEditorPlugin_SuperMarioWorld;
 
-public class SuperMarioWorld : IRetroPlugin , IMenuProvider, ISave
+using SuperNintendoEntertainmentSystem.Memory;
+public class SuperMarioWorld : IRetroPlugin, IMenuProvider, ISave
 {
-	public static string Name => "Super Mario World";
-	public string RomPluginName => "SNES";
-	public bool RequiresAutoLoad => false;
+    public static string Name => "Super Mario World";
+    public string RomPluginName => "SNES";
+    public bool RequiresAutoLoad => false;
 
     byte[] smw_japan = new byte[] { 0x4e, 0x4f, 0x8f, 0x4c, 0xfd, 0xaa, 0xbf, 0xfd, 0xde, 0x20, 0xc8, 0x05, 0x32, 0x02, 0xd4, 0xf0 };
 
@@ -29,8 +30,8 @@ public class SuperMarioWorld : IRetroPlugin , IMenuProvider, ISave
         return hash.SequenceEqual(smw_japan);
     }
 
-	public void SetupGameTemporaryPatches(IMemoryAccess memoryAccess) 
-    { 
+    public void SetupGameTemporaryPatches(IMemoryAccess memoryAccess)
+    {
 
     }
 
@@ -64,25 +65,25 @@ public class SuperMarioWorld : IRetroPlugin , IMenuProvider, ISave
         // Lets attempt a test export
         //
         // Compute checksum of ROM
-        UInt16 checksum=0;
+        UInt16 checksum = 0;
         _exportedData = new ExportMemoryAccess(memoryAccess.RomSize * 2);
         var allBytes = memoryAccess.ReadBytes(ReadKind.Rom, 0, (uint)memoryAccess.RomSize);
         _exportedData.WriteBytes(WriteKind.SerialisedRom, 0, allBytes);
         // Modify header rom size
-        var oldHeaderSize =  _exportedData.ReadBytes(ReadKind.Rom, 0x7FD7, 1)[0];
+        var oldHeaderSize = _exportedData.ReadBytes(ReadKind.Rom, 0x7FD7, 1)[0];
         oldHeaderSize++;
         _exportedData.WriteBytes(WriteKind.SerialisedRom, 0x7FD7, new byte[] { oldHeaderSize });
 
         // Save level modifications
 
         var levelSelect = 199u;
-        var addressTranslation = new LoRom();
+        var addressTranslation = new LoRom(false, false);
         var smwRom = new SuperMarioWorldRomHelpers(memoryAccess, addressTranslation, levelSelect);
 
 
         // Compute checksum and re-record
 
-        for (uint a=0;a<_exportedData.RomSize;a++)
+        for (uint a = 0; a < _exportedData.RomSize; a++)
         {
             checksum += _exportedData.ReadBytes(ReadKind.Rom, a, 1)[0];
         }
@@ -106,30 +107,36 @@ public class SuperMarioWorld : IRetroPlugin , IMenuProvider, ISave
 
     public void ConfigureMenu(IMemoryAccess rom, IMenu menu)
     {
-            menu.AddItem("Level", 
-                (editorInterface,menuItem) => {
-                    editorInterface.OpenUserWindow($"Level View", GetImage(editorInterface, rom));
-                });
-            menu.AddItem("GFX",
-                (editorInterface,menuItem) => {
-                    editorInterface.OpenUserWindow($"GFX View", GetGFXPageImage(editorInterface, rom));
-                });
-            menu.AddItem("VRAM",
-                (editorInterface,menuItem) => {
-                    editorInterface.OpenUserWindow($"VRAM View", GetVRAMImage(editorInterface, rom));
-                });
-            menu.AddItem("Palette",
-                (editorInterface,menuItem) => {
-                    editorInterface.OpenUserWindow($"Palette View", GetPaletteImage(editorInterface, rom));
-                });
-            menu.AddItem("Map 16",
-                (editorInterface,menuItem) => {
-                    editorInterface.OpenUserWindow($"Map16 View", GetMap16Image(editorInterface, rom));
-                });
-            menu.AddItem("Level Editor", 
-                (editorInterface,menuItem) => {
-                    editorInterface.OpenUserWindow($"Level Editor", GetLevelEditor(editorInterface, rom));
-                });
+        menu.AddItem("Level",
+            (editorInterface, menuItem) =>
+            {
+                editorInterface.OpenUserWindow($"Level View", GetImage(editorInterface, rom));
+            });
+        menu.AddItem("GFX",
+            (editorInterface, menuItem) =>
+            {
+                editorInterface.OpenUserWindow($"GFX View", GetGFXPageImage(editorInterface, rom));
+            });
+        menu.AddItem("VRAM",
+            (editorInterface, menuItem) =>
+            {
+                editorInterface.OpenUserWindow($"VRAM View", GetVRAMImage(editorInterface, rom));
+            });
+        menu.AddItem("Palette",
+            (editorInterface, menuItem) =>
+            {
+                editorInterface.OpenUserWindow($"Palette View", GetPaletteImage(editorInterface, rom));
+            });
+        menu.AddItem("Map 16",
+            (editorInterface, menuItem) =>
+            {
+                editorInterface.OpenUserWindow($"Map16 View", GetMap16Image(editorInterface, rom));
+            });
+        menu.AddItem("Level Editor",
+            (editorInterface, menuItem) =>
+            {
+                editorInterface.OpenUserWindow($"Level Editor", GetLevelEditor(editorInterface, rom));
+            });
     }
 
     public SuperMarioWorldLevelViewImage GetImage(IEditor editorInterface, IMemoryAccess rom)
@@ -141,7 +148,7 @@ public class SuperMarioWorld : IRetroPlugin , IMenuProvider, ISave
     {
         return new SuperMarioWorldGFXPageImage(editorInterface, rom);
     }
-    
+
     public SuperMarioWorldVramImage GetVRAMImage(IEditor editorInterface, IMemoryAccess rom)
     {
         return new SuperMarioWorldVramImage(editorInterface, rom);
@@ -151,7 +158,7 @@ public class SuperMarioWorld : IRetroPlugin , IMenuProvider, ISave
     {
         return new SuperMarioWorldPaletteImage(editorInterface, rom);
     }
-    
+
     public SuperMarioWorldMap16Image GetMap16Image(IEditor editorInterface, IMemoryAccess rom)
     {
         return new SuperMarioWorldMap16Image(editorInterface, rom);
