+++ Plugins/Source/GamePlugins/SuperMarioWorld/Views/Map16View.cs
@@ -1,6 +1,8 @@
 using System;
 using RetroEditor.Plugins;
 
+using SuperNintendoEntertainmentSystem.Memory;
+
 namespace RetroEditorPlugin_SuperMarioWorld
 {
     public class SuperMarioWorldMap16Image : IImage, IUserWindow
@@ -26,7 +28,7 @@ namespace RetroEditorPlugin_SuperMarioWorld
         public SuperMarioWorldMap16Image(IEditor editorInterface, IMemoryAccess rom)
         {
             _rom = rom;
-            _addressTranslation = new LoRom();
+            _addressTranslation = new LoRom(false,false);
         }
 
         public void ConfigureWidgets(IMemoryAccess rom, IWidget widget, IPlayerControls playerControls)
