+++ Plugins/Source/GamePlugins/SuperMarioWorld/Views/PaletteView.cs
@@ -2,6 +2,8 @@
 using System;
 using RetroEditor.Plugins;
 
+using SuperNintendoEntertainmentSystem.Memory;
+
 namespace RetroEditorPlugin_SuperMarioWorld
 {
     public class SuperMarioWorldPaletteImage : IImage, IUserWindow
@@ -26,7 +28,7 @@ namespace RetroEditorPlugin_SuperMarioWorld
         public SuperMarioWorldPaletteImage(IEditor editorInterface, IMemoryAccess rom)
         {
             _rom = rom;
-            _addressTranslation = new LoRom();
+            _addressTranslation = new LoRom(false,false);
         }
 
         public void ConfigureWidgets(IMemoryAccess rom, IWidget widget, IPlayerControls playerControls)
