+++ RetroEditorApplication.csproj
@@ -13,6 +13,7 @@
     <Compile Remove="Plugins\**" />
     <Compile Remove="RetroEditor\**" />
     <Compile Remove="Tests\**" />
+    <Compile Remove="SystemPlugins\**" />
     <Content Include="Plugins\Source\GamePlugins\**" CopyToPublishDirectory="PreserveNewest" />
     <Content Include="Plugins\Source\RomPlugins\**" CopyToPublishDirectory="PreserveNewest" />
     <Content Include="Plugins\Plugins.csproj" CopyToPublishDirectory="PreserveNewest" />
