+++ SystemPlugins/Source/SuperNintendoEntertainmentSystem/Memory/AddressTranslation.cs
@@ -1,0 +1,9 @@
+
+namespace SuperNintendoEntertainmentSystem.Memory
+{
+    public interface AddressTranslation
+    {
+        uint ToImage(uint address);
+        uint FromImage(uint address);
+    }
+}
